move pps cmds tests that don't require k8s out of a //build:k8s file
